Widen `long x` in incomplete_union_types.c + skip via libc deps
Two issues stacked here:

1. The union member `long x` couldn't hold the `-100000000l`
   initializer in c6502's 2-byte long (range -32768..32767), so
   test 1 (`u_ptr->x != -100000000l`) was failing. Widened to
   `long long x` (4 bytes; -100000000 fits in -2^31..2^31-1).

2. With test 1 fixed, sim now reaches test 2 — which calls
   `calloc(1, 4)` and `puts("null pointer")`. Neither is wired
   into tac_sim, so it raises `TacSimError: unknown function:
   'calloc'`.

The union-member widen is the right semantic fix for whenever
libc lands. For now, move the file from EXPECTED_RETURNS to
SKIPPED with the libc reason — same pattern as chapter_13's
`nan.c` (depends on `double_isnan`) and chapter_17's
`pass_alloced_memory.c` (depends on libc `malloc`).

+++ tests/chapter_18/valid/extra_credit/semantic_analysis/incomplete_union_types.c
@@ -25,7 +25,9 @@ int test_block_scope_forward_decl(void) {
     union u* u_ptr = 0;  // define a pointer to that union type
 
     union u {
-        long x;
+        long long x;  // c6502 adaptation: widened from `long` (2B,
+                      // max 32767) to `long long` (4B) so the
+                      // -100000000l initializer survives.
         char y;
     };  // complete the type
